+++ client/app/modules/locationmanager/locationeditor.controller.js
@@ -1,36 +1,43 @@
 export class LocationEditorController {
   /*@ngInject*/
-  constructor($uibModalInstance, LocationService, locationSelectedForEditing) {
+  constructor($timeout, $uibModalInstance, LocationService, locationBeforeEdits) {
     // Dependencies
+    this.$timeout = $timeout; // force digest cycle
     this.$uibModalInstance = $uibModalInstance;
     this.locationService = LocationService;
-    this.locationSelectedForEditing = locationSelectedForEditing;
+    this.locationBeforeEdits = locationBeforeEdits;
 
     // Initializations - not in $onInit since not it's own component
     this.submitted = false;
     this.errors = {};
-    this.location = { ...this.locationSelectedForEditing };
+    this.location = { ...this.locationBeforeEdits };
   }
 
   async submitLocation(form) {
     this.submitted = true;
     if(form.$valid) {
-      // Make a copy of this.user or upsert fails
-      let upsertedLocation = { ...this.location };
-      upsertedLocation._id = await this.locationService.locationUpsert(upsertedLocation);
+      try {
+        // Make a copy of this.user or upsert fails
+        this.location._id = await this.locationService.locationUpsert(this.location);
 
-      // Graft the edited location back the original
-      Object.assign(this.locationSelectedForEditing, upsertedLocation);
-      this.$uibModalInstance.close();
+        // Graft the edited location back the original
+        Object.assign(this.locationBeforeEdits, this.location);
+        this.$uibModalInstance.close();
 
-      // Success
-      return true;
+        // Success
+        return true;
+      } catch(err) {
+        this.$timeout(() => {
+          form.locationName.$setValidity('server', false);
+          this.errors.locationName = 'That location name is already being used. Please choose another.';
+        });
+      }
     }
   }
 
   cancel() {
-    if(!this.locationSelectedForEditing._id) {
-      this.locationSelectedForEditing.shouldBeDeleted = true;
+    if(!this.locationBeforeEdits._id) {
+      this.locationBeforeEdits.shouldBeDeleted = true;
     }
     this.$uibModalInstance.close();
   }