+++ client/app/modules/schedulemanager/scheduleeditor.controller.js
@@ -1,10 +1,10 @@
 // Note: if any picklists ever fail to load, may need to have admin component wait for Classes, Teachers, and Location service to initialize
 export class ScheduleEditorController {
   /*@ngInject*/
-  constructor($uibModalInstance, scheduleItemSelectedForEditing, ClassService, TeacherService, LocationService) {
+  constructor($uibModalInstance, scheduleItemBeforeEdits, ClassService, TeacherService, LocationService) {
     // Dependencies
     this.$uibModalInstance = $uibModalInstance;
-    this.scheduleItemSelectedForEditing = scheduleItemSelectedForEditing;
+    this.scheduleItemBeforeEdits = scheduleItemBeforeEdits;
     this.classService = ClassService;
     this.teacherService = TeacherService;
     this.locationService = LocationService;
@@ -12,8 +12,7 @@ export class ScheduleEditorController {
     // Initializations - not in $onInit since not it's own component
     this.weekdays = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
     this.submitted = false;
-    this.errors = {};
-    this.scheduleItem = { ...this.scheduleItemSelectedForEditing };
+    this.scheduleItem = { ...this.scheduleItemBeforeEdits };
     this.options = {
       timeSecondsFormat: 'ss',
       timeStripZeroSeconds: true,
@@ -24,14 +23,11 @@ export class ScheduleEditorController {
   async submitScheduleItem(form) {
     this.submitted = true;
     if(form.$valid) {
-      // Make a copy of this.user or upsert fails
-      let upsertedScheduleItem = { ...this.scheduleItem };
-
       // Set _id to generated one (for inserts) or existing (for updates)
-      upsertedScheduleItem._id = await this.classService.scheduleItemUpsert(upsertedScheduleItem);
+      this.scheduleItem._id = await this.classService.scheduleItemUpsert(this.scheduleItem);
 
       // Graft the edited scheduled item back the original
-      Object.assign(this.scheduleItemSelectedForEditing, upsertedScheduleItem);
+      Object.assign(this.scheduleItemBeforeEdits, this.scheduleItem);
       this.$uibModalInstance.close();
 
       // Successful - return promise
@@ -40,8 +36,8 @@ export class ScheduleEditorController {
   }
 
   cancel() {
-    if(!this.scheduleItemSelectedForEditing._id) {
-      this.scheduleItemSelectedForEditing.shouldBeDeleted = true;
+    if(!this.scheduleItemBeforeEdits._id) {
+      this.scheduleItemBeforeEdits.shouldBeDeleted = true;
     }
     this.$uibModalInstance.close();
   }